backend/requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
python-dotenv
openai>=1.0